+++ packages/website/src/components/Footer/Footer.data.tsx
@@ -1,5 +1,5 @@
 import React, { ReactElement } from 'react';
-import { pageNavigation } from 'src/PageNavigation';
+import { pageNavigation } from 'src/pageNavigation_';
 
 import { SolidFacebookIcon } from '@coderscamp/ui/icons/SolidFacebook';
 import { SolidGitHubIcon } from '@coderscamp/ui/icons/SolidGitHub';
